requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==22.12.0
click==8.1.3
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
iniconfig==2.0.0
ipykernel==6.19.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.5
packaging==22.0
pandas==1.5.2
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg==3.1.4
psycopg-binary==3.1.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
six==1.16.0
stack-data==0.6.2
tokenize-rt==5.0.0
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
wcwidth==0.2.5
